+++ src/playground-ide.ts
@@ -242,7 +242,7 @@ export class PlaygroundIde extends LitElement {
   /**
    * How to handle `playground-hide` and `playground-fold` comments.
    *
-   * See https://github.com/PolymerLabs/playground-elements#hiding--folding for
+   * See https://github.com/google/playground-elements#hiding--folding for
    * more details.
    *
    * Options: